chore: 팀 코드리뷰 변경사항 반영

+++ matjalalzz/src/main/java/shop/matjalalzz/tosspay/dto/TossPaymentConfirmRequest.java
@@ -4,9 +4,10 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public record TossPaymentConfirmRequest(
-	@NotBlank(message = "결제 키(paymentKey)는 필수 입력값입니다.") String paymentKey,
+    @NotBlank(message = "결제 키(paymentKey)는 필수 입력값입니다.") String paymentKey,
 
-	@NotBlank(message = "주문 번호(orderId)는 필수 입력값입니다.") String orderId,
+    @NotBlank(message = "주문 번호(orderId)는 필수 입력값입니다.") String orderId,
 
-	@NotNull(message = "결제 키(paymentKey)는 필수 입력값입니다.") Long amount) {
+    @NotNull(message = "결제 키(paymentKey)는 필수 입력값입니다.") int amount) {
+
 }
